Initial Commit.
This is the first data set I have uploaded in a couple of months.
In terms of days, it contains basically an additional month's worth of
data.

However, there were days when the Nightscout rig did not operate because
of travel and other reasons. For example, for several days, we ran out
of sensors in the middle of August, so there is a decent sized gap
there.

Oh well.

 Changes to be committed:
	new file:	data/entries-2015-09-01.csv