Initial Commit
This file is a heavily documented, but simple script I can use to
quickly create new Nightscout CGM data sets. I don't need to re-compile
the get-cgm-data.Rmd file. This works just as well, in fact it is based
on the code found in that file, but doesn't need to create and
side-affects other than creating a new data set.

 Changes to be committed:
	new file:	get-nightscout-data-from-server.R